Trying to Prevent AppVeyor Builds from Searching for LaTeX in Tests
The AppVeyor builds all time out. Maybe it is because they are trying
to look for a LaTeX installation. If that is true, then we may make
them feasible by switching off this behavior.

The new tests do not compile LaTeX documents if a system property
"dontCompileLaTeX" is set. We set it via the new "passToMaven" command
line argument of our Ant build script.